+++ samples-lib/src/main/java/com/esri/arcgismaps/sample/sampleslib/components/adaptive/AdaptiveThreePane.kt
@@ -395,3 +395,139 @@ private const val AUTO_CLOSE_EXPANSION_ANCHOR_INDEX = 3
 
 // Auto-close threshold when supporting pane shrinks to a thin strip.
 private const val WIDE_CLOSE_PROPORTION_THRESHOLD = 0.80f
+
+@Preview(
+    name = "Pane preview - main only day",
+    widthDp = 411,
+    heightDp = 891,
+    showBackground = true,
+)
+@Preview(
+    name = "Pane preview - main only night",
+    widthDp = 411,
+    heightDp = 891,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+private fun AdaptiveThreePaneMainOnlyPreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(supportingPaneInitiallyOpen = false),
+    )
+}
+
+@Preview(
+    name = "Pane preview - supporting open",
+    widthDp = 411,
+    heightDp = 891,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+private fun AdaptiveThreePaneSupportingPanePreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(supportingPaneInitiallyOpen = true),
+    )
+}
+
+@Preview(
+    name = "Pane preview - floating visible",
+    widthDp = 891,
+    heightDp = 411,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES,
+)
+@Composable
+private fun AdaptiveThreePaneFloatingPanePreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(
+            supportingPaneInitiallyOpen = true,
+            floatingPaneInitiallyVisible = true
+        ),
+    )
+}
+
+@Preview(
+    name = "Phone portrait - compact 50%",
+    widthDp = 411,
+    heightDp = 891,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+private fun AdaptiveThreePanePhonePortraitPreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(
+            supportingPaneInitiallyOpen = true,
+            compactSupportingPaneHeightRatio = 0.5f,
+        ),
+    )
+}
+
+@Preview(
+    name = "Phone landscape - compact 35%",
+    widthDp = 891,
+    heightDp = 411,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES,
+)
+@Composable
+private fun AdaptiveThreePanePhoneLandscapePreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(
+            supportingPaneInitiallyOpen = true,
+            compactSupportingPaneHeightRatio = 0.35f
+        ),
+    )
+}
+
+@Preview(
+    name = "Tablet portrait - supporting open",
+    widthDp = 800,
+    heightDp = 1280,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+private fun AdaptiveThreePaneTabletPortraitPreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(
+            supportingPaneInitiallyOpen = true,
+            compactSupportingPaneHeightRatio = 0.6f
+        ),
+    )
+}
+
+@Preview(
+    name = "Tablet landscape - floating visible",
+    widthDp = 1280,
+    heightDp = 800,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+private fun AdaptiveThreePaneTabletLandscapePreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(
+            supportingPaneInitiallyOpen = true,
+            floatingPaneInitiallyVisible = true
+        ),
+    )
+}
+
+@Preview(
+    name = "Compact ratio 65%",
+    widthDp = 411,
+    heightDp = 891,
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+private fun AdaptiveThreePaneCompactRatioPreview() {
+    AdaptiveThreePanePreviewContent(
+        config = ThreePaneConfig(
+            supportingPaneInitiallyOpen = true,
+            compactSupportingPaneHeightRatio = 0.65f
+        )
+    )
+}
