+++ samples/show-contrast-responsive-geoview/src/main/java/com/esri/arcgismaps/sample/showcontrastresponsivegeoview/MainActivity.kt
@@ -19,9 +19,9 @@ package com.esri.arcgismaps.sample.showcontrastresponsivegeoview
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
 import com.arcgismaps.ApiKey
 import com.arcgismaps.ArcGISEnvironment
 import com.esri.arcgismaps.sample.sampleslib.theme.SampleAppTheme
@@ -34,7 +34,7 @@ class MainActivity : ComponentActivity() {
         // authentication with an API key or named user is
         // required to access basemaps and other location services
         ArcGISEnvironment.apiKey = ApiKey.create(BuildConfig.ACCESS_TOKEN)
-
+        enableEdgeToEdge()
         setContent {
             SampleAppTheme {
                 Surface(color = MaterialTheme.colorScheme.background) {
